+++ src/app/profile/profile.page.ts
@@ -246,7 +246,7 @@ export class ProfilePage implements OnInit {
       this.ui.presentToast('Locations now enabled', this.toastController);
       environment.overrideLocations = true;
       this.settings.save();
-      this.db.setLocationAvailable({ art: false, camps: false, artMessage: '', campMessage: '' });
+      this.db.setLocationHidden({ art: false, camps: false, artMessage: '', campMessage: '' });
       await this.db.populate(this.settings.settings.datasetId, this.db.getTimeZone());
       this.db.resume.set(new Date().toString());
     }