+++ src/app/tabs/tabs.page.ts
@@ -81,7 +81,7 @@ export class TabsPage implements OnInit {
       const timeZone = this.db.getTimeZone();
       if (this.db.locationsHidden() && !hide) {
         // Locations were unlocked
-        this.db.setLocationAvailable({ art: false, camps: false, artMessage: '', campMessage: '' });
+        this.db.setLocationHidden({ art: false, camps: false, artMessage: '', campMessage: '' });
         await this.db.populate(this.settings.settings.datasetId, timeZone);
       }
       this.ui.setStatusBarBasedOnTheme();